Create README.md
Description

Sousmate is a kitchen aid, helping users to cook their favorite recipes. Recipes can be uploaded by Users, which Sousmate will display for users cooking that recipe in a step-by-step fashion. Sousmate will track the cooking time for foods, and will alert users when certain periods of time have elapsed.

User Roles:

All:
create a recipe
view recipe
 

Primary stakeholders have decided that Sousmate should have the ability to see a count-down clock for a chosen recipe. The development team is charged with implementing such a feature.

 

----

The next sections are the Purpose, Scope, and System Overview. These sections explain the purpose of the document, the overal scope of the document (what's involved), and the architecture of the system, respectively. 

The Purpose, Scope, and System Overview for Sousmate are: 

Purpose

The purpose of this document is to provide a guide for developers and testers who are responsible for the development of the platform. It provides the information necessary to design, develop, and test the software.

 

Scope

This document serves as a description of the required improvements to the project. It contains a workflow description, functional and nonfunctional requirements, and appropriate diagrams to fully describe the system.

 

System Overview

This enterprise project consists of one layer. The front-end is the one layer and is the portal through which users will interact with the application. Primary interactivity is achieved using JavaScript executing in the browser.

 

<hr/>

The last sections are reserved for UML (Unified Modeling Language) diagrams. These diagrams convey different functions of the system, how users interact with the system, and such. 

One particular diagram is a Use Case Diagram. It depicts how users interact with the system. 

The Use Cases for Foodar are below. In a future Dive, you will make your own Use Case diagram.

 

All Users
Recipe System
Recipe CRUD Operations
Create a Recipe from the web user interface
Read a recipe
Make changes to a current recipe
remove recipe from list
A timer for users to watch time count down from total cook-time to zero